Fix library structure and some typo name
